wrapper script for migrations and og setup script
